Move normalize_uri, normalize_parsed_uri to separate mod
