+++ nexus-common/src/universal_tag/mod.rs
@@ -1,2 +1,3 @@
 pub mod app_tag_info;
 pub mod homeserver_parsed_uri;
+pub mod normalize;
